fix: Deno /db/complex — cast numeric to float8 (driver returns string)
deno-postgres returns DECIMAL columns as strings by default. The
JSON.stringify of {totalValue: "1234.56"} doesn't crash, but oak's
response serializer may. Cast COUNT to int and SUM/AVG to float8 in
the SQL to avoid the numeric-as-string issue.

Same pattern in all 3 Deno impls (oak, deno_serve, fresh).

+++ src/deno/oak/services/database.ts
@@ -65,9 +65,9 @@ export class DatabaseService {
       SELECT
         u.id AS "userId",
         u.first_name || ' ' || u.last_name AS "userName",
-        COUNT(o.id) AS "totalOrders",
-        COALESCE(SUM(o.total_amount), 0) AS "totalValue",
-        COALESCE(AVG(o.total_amount), 0) AS "averageOrderValue"
+        COUNT(o.id)::int AS "totalOrders",
+        COALESCE(SUM(o.total_amount), 0)::float8 AS "totalValue",
+        COALESCE(AVG(o.total_amount), 0)::float8 AS "averageOrderValue"
       FROM users u
       INNER JOIN orders o ON u.id = o.user_id
         WHERE o.created_at >= NOW() - INTERVAL '1 day' * $1
